requirements.txt
# Core FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0

# Image processing and computer vision
Pillow==10.1.0
numpy==1.26.4
opencv-python-headless==4.8.1.78

# YOLOv12 and PyTorch (compatible versions for ONNX export)
ultralytics>=8.3.172
onnx>=1.15.0

# OpenVINO for optimized inference
openvino>=2023.1.0

# System and performance monitoring
py-cpuinfo==9.0.0

# Additional dependencies for complete functionality
requests>=2.31.0
tqdm>=4.64.0
aiohttp>=3.8.0
redis>=5.0
# httpx pinned below 0.27 for compatibility with starlette TestClient
httpx<0.27

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Performance monitoring
psutil>=5.9.0

